Improve C++ GitHub pipelines (#219)
* change sonar property

Signed-off-by: Nicolas Rol <nicolas.rol@rte-france.com>

* copy boost headers in fork case

Signed-off-by: Nicolas Rol <nicolas.rol@rte-france.com>

* fix branch in git diff

Signed-off-by: Nicolas Rol <nicolas.rol@rte-france.com>

* Set fetch-depth to 0 in checkout step of dev CI workflow

This ensures the full history is fetched during the checkout process. It is necessary for workflows that rely on complete commit history, such as versioning or changelog generation.

Signed-off-by: Nicolas Rol <nicolas.rol@rte-france.com>

* Update SonarScanner version in CI workflow

Bump SonarScanner version from 3.3.0.1492 to 7.1.0.4889 in the dev CI pipeline. This ensures compatibility with newer scanning capabilities and potential bug fixes in the updated version.

Signed-off-by: Nicolas Rol <nicolas.rol@rte-france.com>

---------

Signed-off-by: Nicolas Rol <nicolas.rol@rte-france.com>